Ref: extracting class Diagnostics.
